Avoid repeated indirect access to dictionaries and multi-dimensional vectors
This is a performance optimization patch that moves complex element
lookups out of the hot loops in onesource::generatePDF(), holding
references (aliases) to all the required objects through the loops.
A variant of this patch, which also included aliases to 1-D vectors
(just for debugging), resulted in an improvement of 12% in the total
execution time of a 24+ hour estimation run with pz-rail-lephare of the
LSST DP0.2 dataset (from 28 hours to 25 hours).

The detailed changes in this patch are the following:

- Remove unordered map lookups from the main onesource::generatePDF()
  loop.
- Move the indirect access of 2 out of the 3 vector dimensions outside a
  nested loop in onesource::generatePDF().
- Replace an explicit loop by an implicit one to copy temporary vectors
  to the final ones.