Update test_suite.sh to run self contained
It previoulsy required data to already be in place so could cause confusion. I also tried to
add some comments to aid the reading of the file. It should now run out of the box following a pip install lephare.
It could still be moved or added as a download in the documentation possibly?