Updates to intermediate notebook following Oliver comments in issue
Some of the change requests were ignored as they are required for the notebook to run.

I have so far left in the onesource examples at the end as they were previously requested. We can discuss if to remove them this week.

The issue with curl to download the para file was that it needed the url updating to the raw file not the html.

The example para file does not run through the downloader because it needs the sed list links to be relative. We can either update the
para file or the downloader to append these links